+++ docker/core/mkmediascanner/src/mk_lib_database_media_tv.rs
@@ -4,11 +4,11 @@
 mod mk_lib_logging;
 
 use serde::{Deserialize, Serialize};
+use serde_json::json;
 use sqlx::postgres::PgRow;
 use sqlx::{types::Json, types::Uuid};
 use sqlx::{FromRow, Row};
 use stdext::function_name;
-use serde_json::json;
 
 #[derive(Debug, FromRow, Deserialize, Serialize)]
 pub struct DBMediaTVShowList {
@@ -21,8 +21,8 @@ pub struct DBMediaTVShowList {
 pub async fn mk_lib_database_media_tv_read(
     sqlx_pool: &sqlx::PgPool,
     search_value: String,
-    offset: i32,
-    limit: i32,
+    offset: i64,
+    limit: i64,
 ) -> Result<Vec<DBMediaTVShowList>, sqlx::Error> {
     #[cfg(debug_assertions)]
     {
