+++ docker/core/mkwebaxum/src/mk_lib_database_metadata_book.rs
@@ -4,11 +4,11 @@
 mod mk_lib_logging;
 
 use serde::{Deserialize, Serialize};
+use serde_json::json;
 use sqlx::postgres::PgRow;
 use sqlx::{types::Json, types::Uuid};
 use sqlx::{FromRow, Row};
 use stdext::function_name;
-use serde_json::json;
 
 pub async fn mk_lib_database_metadata_book_detail(
     sqlx_pool: &sqlx::PgPool,
@@ -42,8 +42,8 @@ pub struct DBMetaBookList {
 pub async fn mk_lib_database_metadata_book_read(
     sqlx_pool: &sqlx::PgPool,
     search_value: String,
-    offset: i32,
-    limit: i32,
+    offset: i64,
+    limit: i64,
 ) -> Result<Vec<DBMetaBookList>, sqlx::Error> {
     #[cfg(debug_assertions)]
     {
